+++ packages/app/public/loading/index.html
@@ -43,10 +43,10 @@
     <h3 class="text dot-animation" id="text">Lade</h3>
   </div>
   <script>
-    let { ipcRenderer } = require('electron')
-    ipcRenderer.on('msg', ($event, msg) => {
-      document.getElementById('text').innerText = msg
-    })
+    // let { ipcRenderer } = require('electron')
+    // ipcRenderer.on('msg', ($event, msg) => {
+    //   document.getElementById('text').innerText = msg
+    // })
   </script>
 </body>
