+++ packages/app/src/lib/import.ts
@@ -3,6 +3,7 @@ import Vue from 'vue';
 import wrapper from './wrapper.lib.vue';
 import suche from './search.lib.vue';
 import editAK from './editAK.form.lib.vue';
+import abmelden from './abmelden.form.lib.vue';
 import editBemerkungen from './editAnmeldung.Bemerkungen.lib.vue';
 // import loadData from './loadData.lib.vue';
 
@@ -18,3 +19,4 @@ Vue.component('ec-lesezeichen-show', showLesezeichen);
 // Vue.component('ec-loadData', loadData);
 Vue.component('ec-form-edit-ak', editAK);
 Vue.component('ec-edit-anmeldung-bemerkungen', editBemerkungen);
+Vue.component('ec-abmelden', abmelden);
